+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Notes about creating components that delegate portal behavior and render generic content provided by parent components.
+
 ## [0.8.0] - 2025-05-03
 
 ### Added